Refactor folder structure and export db files from database folder
