+++ database/index.js
@@ -1,0 +1,4 @@
+const usersDatabase = require("./users-database");
+const createdRoomsDatabase = require("./createdRooms-database");
+
+module.exports = { usersDatabase, createdRoomsDatabase };
